+++ realm/realm-library/src/objectServer/java/io/realm/permissions/PermissionChange.java
@@ -23,7 +23,9 @@ import io.realm.annotations.PrimaryKey;
 import io.realm.annotations.Required;
 
 /**
- * TODO
+ * This class is used for requesting changes to a Realms permissions.
+ *
+ * @see <a href="`https://realm.io/docs/realm-object-server/#permissions">Controlling Permissions</a>
  */
 public class PermissionChange extends RealmObject {
 
@@ -42,22 +44,73 @@ public class PermissionChange extends RealmObject {
     private String realmUrl;
     @Required
     private String userId;
-    private boolean mayRead;
-    private boolean mayWrite;
-    private boolean mayManage;
+    private Boolean mayRead;
+    private Boolean mayWrite;
+    private Boolean mayManage;
 
     public PermissionChange() {
         // Default constructor required by Realm
     }
 
     /**
-     * TODO
+     * Construct a Permission Change Object.
+     *
+     * @param realmUrl Realm to change permissions for.
+     * @param userId User or users to effect.
+     * @param mayRead Control read access. {@code true} or {@code false} to request this new value. {@code null} to
+     *                keep current value.
+     * @param mayWrite Control write access. {@code true} or {@code false} to request this new value. {@code null} to
+     *                 keep current value.
+     * @param mayManage Control manage access. {@code true} or {@code false} to request this new value. {@code null} to
+     *                  keep current value.
+     *
+     * @see <a href="`https://realm.io/docs/realm-object-server/#permissions">Controlling Permissions</a>
      */
-    public PermissionChange(String realmUrl, String userId, boolean mayRead, boolean mayWrite, boolean mayManage) {
+    public PermissionChange(String realmUrl, String userId, Boolean mayRead, Boolean mayWrite, Boolean mayManage) {
         this.realmUrl = realmUrl;
         this.userId = userId;
         this.mayRead = mayRead;
         this.mayWrite = mayWrite;
         this.mayManage = mayManage;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public Date getCreatedAt() {
+        return createdAt;
+    }
+
+    public Date getUpdatedAt() {
+        return updatedAt;
+    }
+
+    public Integer getStatusCode() {
+        return statusCode;
+    }
+
+    public String getStatusMessage() {
+        return statusMessage;
+    }
+
+    public String getRealmUrl() {
+        return realmUrl;
+    }
+
+    public String getUserId() {
+        return userId;
+    }
+
+    public Boolean mayRead() {
+        return mayRead;
+    }
+
+    public Boolean mayWrite() {
+        return mayWrite;
+    }
+
+    public Boolean mayManage() {
+        return mayManage;
     }
 }